feat: add `--binary-path` to `run-ios` and `run-android` to enable installing pre-build binaries (#1)
* feat: add --binary-path support to run-ios

* refactor: extract tryInstallAppOnDevice

* feat: add --binary-path support to run-android

* lint

* refactor: support absolute binaryPath
