+++ packages/platform-android/src/commands/runAndroid/tryInstallAppOnDevice.ts
@@ -1,0 +1,72 @@
+import execa from 'execa';
+import fs from 'fs';
+import {logger, CLIError} from '@react-native-community/cli-tools';
+
+import adb from './adb';
+import type {AndroidProject, Flags} from './';
+
+function tryInstallAppOnDevice(
+  args: Flags,
+  adbPath: string,
+  device: string,
+  androidProject: AndroidProject,
+) {
+  try {
+    // "app" is usually the default value for Android apps with only 1 app
+    const {appName, sourceDir} = androidProject;
+    const {appFolder} = args;
+    const variant = args.variant.toLowerCase();
+
+    let pathToApk;
+    if (!args.binaryPath) {
+      const buildDirectory = `${sourceDir}/${appName}/build/outputs/apk/${variant}`;
+      const apkFile = getInstallApkName(
+        appFolder || appName, // TODO: remove appFolder
+        adbPath,
+        variant,
+        device,
+        buildDirectory,
+      );
+      pathToApk = `${buildDirectory}/${apkFile}`;
+    } else {
+      pathToApk = args.binaryPath;
+    }
+
+    const adbArgs = ['-s', device, 'install', '-r', '-d', pathToApk];
+    logger.info(`Installing the app on the device "${device}"...`);
+    logger.debug(
+      `Running command "cd android && adb -s ${device} install -r -d ${pathToApk}"`,
+    );
+    execa.sync(adbPath, adbArgs, {stdio: 'inherit'});
+  } catch (error) {
+    throw new CLIError('Failed to install the app on the device.', error);
+  }
+}
+
+function getInstallApkName(
+  appName: string,
+  adbPath: string,
+  variant: string,
+  device: string,
+  buildDirectory: string,
+) {
+  const availableCPUs = adb.getAvailableCPUs(adbPath, device);
+
+  // check if there is an apk file like app-armeabi-v7a-debug.apk
+  for (const availableCPU of availableCPUs.concat('universal')) {
+    const apkName = `${appName}-${availableCPU}-${variant}.apk`;
+    if (fs.existsSync(`${buildDirectory}/${apkName}`)) {
+      return apkName;
+    }
+  }
+
+  // check if there is a default file like app-debug.apk
+  const apkName = `${appName}-${variant}.apk`;
+  if (fs.existsSync(`${buildDirectory}/${apkName}`)) {
+    return apkName;
+  }
+
+  throw new CLIError('Could not find the correct install APK file.');
+}
+
+export default tryInstallAppOnDevice;
